update:优化了输入上传的带宽，60tick下从50KB/s变成了10KB/s，可喜可贺

+++ Assets/StargateNet/StargateNet/Base/SimulationInput.cs
@@ -10,7 +10,7 @@ namespace StargateNet
     {
         public struct InputBlock
         {
-            public int type;
+            public short type;
             public INetworkInput input;
         }
 
@@ -33,7 +33,7 @@ namespace StargateNet
             inputBlocks.Add(inputBlock);
         }
 
-        public void AddInputBlock(int type, INetworkInput input)
+        public void AddInputBlock(short type, INetworkInput input)
         {
             this.inputBlocks.Add(new InputBlock { type = type, input = input });
         }
